excluded uuid from label of rescaled structure
