+++ aiida_fleur/tests/workflows/test_create_magnetic_workchain.py
@@ -34,7 +34,7 @@ class Test_FleurCreateMagneticWorkChain:
     Regression tests for the FleurCreateMagneticWorkChain
     """
     @pytest.mark.timeout(500, method='thread')
-    def test_fleur_create_mag_FePt(self, run_with_cache, fleur_local_code, inpgen_local_code, clear_database):
+    def test_fleur_create_mag_FePt(self, run_with_cache, fleur_local_code, inpgen_local_code, clear_spec):
         """
         full example using scf workflow with just a fleurinp data as input.
         Several fleur runs needed till convergence
